feat(product-service): refactor product management with new DTOs, service, and controller structure; implement pagination and response mapping update(user-service-application-properties): to match the api-gateway service name

+++ product-service/src/main/java/com/buy01/product/DTOs/UpdateRequest.java
@@ -1,5 +1,7 @@
 package com.buy01.product.DTOs;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -7,16 +9,11 @@ import jakarta.validation.constraints.Size;
 
 public record UpdateRequest(
 
-    @NotBlank
-    @Size(min = 3, max = 100)
-    String name,
+        @NotBlank @Size(min = 3, max = 100) String name,
 
-    @NotBlank
-    @Size(min = 3, max = 1000)
-    String description,
+        @NotBlank @Size(min = 3, max = 1000) String description,
 
-    @NotNull
-    @DecimalMin(value = "0.0", inclusive = false)
-    Double price
+        @NotNull @DecimalMin(value = "0.0", inclusive = false) BigDecimal price
 
-) {}
+) {
+}
